refactor(router): add app routes to the default export
